优化如下代码： 1. 移除Question、Blog、Comment、Likes类的id0数据，增加vo类 2. 新增CopyVo方法，剩去麻烦的set get 3. 评论和点赞归档新增动作接收者（问题或博客）的id 4. 优化搜索逻辑，仅搜索公开和没有隐藏的问题和博客 修复如下bug： 1. 查询当前用户有没有对某博客/问题点过赞时，不再查询用户对其下评论的点赞记录

+++ src/main/java/com/lrm/web/admin/TagController.java
@@ -83,6 +83,9 @@ public class TagController {
             throw new NotFoundException("未查询到该标签");
         }
 
+        if (tag.getQuestions() != null || tag.getBlogs() != null) {
+            throw new FailedOperationException("有博客或问题使用了该标签，无法删除");
+        }
         tagServiceImpl.deleteTag(tagId);
         tag = tagServiceImpl.getTag(tagId);
         if (tag != null) {
